ORCH-0964: keep mobile-web bottom sheet (fix recursion via neutral SheetMobile)
Supersedes the centred-card approach from the previous commit: phone-width
web again renders the canonical bottom sheet, with the recursion root-caused
and fixed properly.

The recursion: Sheet.web.tsx imported the Sheet value from "./Sheet". On web
Metro resolves "./Sheet" to Sheet.web.tsx itself, so the narrow-web branch
rendered itself with no base case -> unbounded fiber tree -> mobile-web
renderer OOM.

Fix: move the bottom-sheet implementation to the platform-NEUTRAL
SheetMobile.tsx (no .web variant, so it never self-resolves). Sheet.tsx (native
entry) re-exports it; Sheet.web.tsx imports MobileSheet from "./SheetMobile" and
keeps the useResponsiveLayout gate (narrow -> bottom sheet, wide -> centred
card). Native behaviour unchanged.

Verified on a real device (Galaxy A72, dpr 2.8): brand page renders + holds at
~190MB; tapping Share mounts the bottom sheet (nodes 85->183, "Copy link"
visible) with no crash. Previously the hidden ShareModal OOM-killed the tab on
page load.

[TEST-MOD-APPROVED ORCH-0964] Sheet.web.test.ts rewritten to anchor the
recursion guard (bottom sheet imported from ./SheetMobile, never a value import
from ./Sheet). KeyboardRoot.test.tsx T-12/13/14 repointed from Sheet.tsx to
SheetMobile.tsx (the impl moved there; Sheet.tsx is now a thin re-export).
